code changes for the sorting and pagination for compleated visits
